Move provider code from /pages/api/providers to /lib/providers
