+++ lib/providers/bandcamp_new.ts
