+++ lib/providers/providers.ts
@@ -4,12 +4,11 @@ import musixmatch from "./musixmatch";
 import deezer from "./deezer";
 import tidal from "./tidal"
 import bandcamp from "./bandcamp"
-import logger from "../../../utils/logger"
 import soundcloud from "./soundcloud"
 import applemusic from "./applemusic";
-import { FullProvider, PartialProvider, Provider, ProviderCapability, ProviderNamespace, ProviderWithCapabilities, UrlInfo } from "../../../types/provider-types";
-import clientProviders from "../../../utils/clientProviders";
-import parsers from "../../../lib/parsers/parsers";
+import { PartialProvider, Provider, ProviderCapability, ProviderNamespace, ProviderWithCapabilities } from "../../types/provider-types";
+import clientProviders from "../../utils/clientProviders";
+import parsers from "../parsers/parsers";
 const { isDisabled } = clientProviders;
 
 const providerList = [
