+++ pages/api/compareArtistAlbums.ts
@@ -1,4 +1,4 @@
-import musicbrainz from "./providers/musicbrainz";
+import musicbrainz from "../../lib/providers/musicbrainz";
 import processData from "../../utils/processAlbumData";
 import logger from "../../utils/logger";
 import { NextApiRequest, NextApiResponse } from "next";
@@ -6,7 +6,7 @@ import { AlbumData, AlbumObject, ExtendedAlbumObject, ProviderWithCapabilities, 
 import { IUrl } from "musicbrainz-api";
 import normalizeVars from "../../utils/normalizeVars";
 import { SAMBLApiError } from "../../types/api-types";
-import providers from "./providers/providers";
+import providers from "../../lib/providers/providers";
 
 // spotifyId - Spotify artist ID
 // mbid - MusicBrainz artist ID. Only neccesary if you want to check if the associated albums are linked to that artist
