+++ pages/api/compareSingleAlbum.ts
@@ -1,6 +1,6 @@
 import logger from "../../utils/logger";
-import providers from "./providers/providers"
-import musicbrainz from "./providers/musicbrainz";
+import providers from "../../lib/providers/providers"
+import musicbrainz from "../../lib/providers/musicbrainz";
 import processData from "../../utils/processAlbumData";
 import { NextApiRequest, NextApiResponse } from "next";
 import normalizeVars from "../../utils/normalizeVars";
