+++ pages/api/searchArtists.ts
@@ -1,5 +1,5 @@
-import musicbrainz from "./providers/musicbrainz";
-import providers from "./providers/providers";
+import musicbrainz from "../../lib/providers/musicbrainz";
+import providers from "../../lib/providers/providers";
 import logger from "../../utils/logger";
 import { NextApiRequest, NextApiResponse } from "next";
 import normalizeVars from "../../utils/normalizeVars";
